Allow toggling without pointer lock acquired - listen to DOM keydown event (vs voxel-keys)
voxel-keys is intended for keybindings in use while interacting with the
game, so it only enables the event listeners when the user has
pointerlock. However, it is still desirable to allow toggling fullscreen
when at GUI elements, e.g., a voxel-workbench. To solve this, listen to
the DOM keydown event ourselves, but still use voxel-keys to lookup the
binding value (in case it is changed with e.g., kb-bindings-ui).